pyproject.toml
[tool.poetry]
name = "lending-app"
version = "0.1.0"
description = "Lending Management App with FastAPI and Supabase"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.29"}
sqlmodel = "^0.0.18"
asyncpg = "^0.29"
alembic = "^1.13"
python-dotenv = "^1.0"
apscheduler = "^3.10"
pydantic-settings = "^2.1.0"
greenlet = "^3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
httpx = "^0.26"
black = "^23.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 